'Home' page example
Please watch the tutorial linked below for working with Yelp Fusion in python. You will be required to create a Yelp account in order to get an API access key.
In order to run the site locally, go to the bottom of views.py, replace 'YOUR_API_KEY_HERE' with your API key, and run the main.py file.

Yelp API Python Tutorial:
https://youtu.be/oggMBtza80E